change variable signedness

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@ void print_number(int n)
 {
 	unsigned int m; /* power of 10 */
 	int c; /* boolean check */
-	unsigned int num; /* convert int to long */
+	long int num; /* convert int to long */
 
 	num = n;
 	/* negatives */
